Cache performance fixes + Vector.toArray allocations
What was happening was that the fuzzy CSG lookup is performance critical, and was using map, forEach, and in instead of more basic javascript operators.  This was doubling the time for rendering an 80 face sphere.

Memory profiling indicated that Vector.toArray is allocating two arrays on every call.  This is because it's calling [ x, y, z].slice(0, n||3) - which allocates the first array, then trims it to the size you want.  Most the time we want all 3 in the array so I have optimized for that and used a switch statement for the less frequently used methods.