Allow multiple workers, but cleanup worker thread after finishing
It does look like the original intent of the code was to have one worker, but this worker was leaking if you never called abort().

The intent here is to have the worker clean itself when it has finished doing its callback.

I tried one worker, but for our purposes it's handier to have multiple workers going.  We can review and kill off old workers by grabbing the builder after setjscad and calling cancel on older workers we might not want to keep going.