Add support for the screenshot button
In order to do this we grab the webgl context out and paint it into a 2d canvas.  We have to turn on preserveDrawingBuffer, otherwise the you'll get a blank canvas.  NOTE: This may lead to a slight degrade in WebGL perf and should be monitored.

This is A LOT quicker than doing .toBlob(), creating an object url, loading into an image and then drawing the image into the 2d canvas.  That process was getting killed by the 1s timeout in pxt screenshot taker.  Both processes require the special preserveDrawingBuffer on the webgl drawing context.